+++ astropy/extern/setup_package.py
@@ -5,7 +5,7 @@ import os
 
 
 def get_package_data():
-    paths = [os.path.join('js', '*.js')]
+    paths = [os.path.join('js', '*.js'), os.path.join('css', '*.css')]
     return {'astropy.extern': paths}